Create Get-WindowsAutopilotInfo.ps1
This script uses WMI to retrieve properties needed for a customer to register a device with Windows Autopilot.  Note that it is normal for the resulting CSV file to not collect a Windows Product ID (PKID) value since this is not required to register a device.  Only the serial number and hardware hash will be populated.

Signed-off-by: DukeDexter <86385353+DukeDexter@users.noreply.github.com>